# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInquiryInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 6
     * @maxLength 48
     */
  phone: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  company: string;
  /** @maxLength 24 */
  date?: string;
  /** @maxLength 32 */
  time?: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  meetingType: string;
  /**
     * @minLength 30
     * @maxLength 6000
     */
  projectSummary: string;
}

export interface NewsletterSubscriptionInput {
  /** @maxLength 255 */
  email: string;
}

export interface SubmissionStatus {
  ok: boolean;
  message: string;
}